I have attempted the first refactoring. see  https://twiki.cern.ch/twiki/bin/view/CMS/PatLayer1Refactoring
The tag before refactoring is Before_Producer_Refactoring

I had to modify bit the cfg/cff files
 - data/PATHighLevelReco = things to happen between cleaners & producers
 - data/PATBeforeLayer0Reco  = things to happen before cleaners
 - data/Layer0: include the two above, define cleaners
 - test/setupForFamos = replaces to run on fastsim  (should go in "/data/" ...)
 - test/boostrapWithFamos= replaces to run on fastsim : gen+fastsim cff file

I have tested it only on a pre-made 168 FastSim sample of 100 ttbar
 test/in.root -> /afs/cern.ch/cms/PRS/top/cmssw-data/csa07/CMSSW_1_6_8/src/ttbar.root
it runs but it complains for some missing dictionary when saving.

taus are disabled, as the RecoTau tags we use conflict with some cff files in release

some things will require the conversion of existing associations to valuemap before layer 0 cleaners + re-keying after + changing the code which reads them in the producers.